+++ ui/src/main/java/com/drone/navdata/DroneNavData.java
@@ -1,15 +1,17 @@
 package com.drone.navdata;
 
-public interface  DroneNavData {
+public interface DroneNavData {
 
-	public double getBattery();
-	
-	public float getPitch();
-	
-	public float getRoll();
-	
-	public float getGaz();
-	
-	public float getYaw();
-	
+    /**
+     * @return Battery level of drone [0 - 100]
+     */
+    public int getBattery();
+
+    public float getPitch();
+
+    public float getRoll();
+
+    public float getGaz();
+
+    public float getYaw();
 }
